updated dep version & asm syntax, binding now OK

+++ apps/synthetic/src/main.rs
@@ -216,7 +216,7 @@ fn run_spawner_server(addr: SocketAddrV4, worker: FakeWorker) {
     }
     extern "C" fn echo(d: *mut shenango::ffi::udp_spawn_data) {
         unsafe {
-            let buf = slice::from_raw_parts((*d).buf as *mut u8, (*d).len);
+            let buf = slice::from_raw_parts((*d).buf as *mut u8, (*d).len as usize);
             let payload = Payload::deserialize(&mut &buf[..]).unwrap();
             let worker = SPAWNER_WORKER.as_ref().unwrap();
             worker.work(payload.work_iterations);
